Main.OurModuleName.MP — Method

MP(y, z)

Compute the middle product of polynomials $y$ and $z$.

For vectors $y = [y_1,...,y_n]$ and $z = [z_1,...,z_{2n-1}]$ of polynomial coefficients, return $[c_{n-1},..., c_{2n-1}]$, where $c_k = \sum_{i + j = k} a_i b_j$.

Examples

julia> Main.OurModuleName.MP([4, 6, 38, -1], [1, 2, 3, 4, 8, 9, 26, 7])
4-element Vector{Int64}:
 109
 168
 233
 458

julia> Main.OurModuleName.MP([-7//10, 1, -2], [9//4, -1//4, -2//5, 9//8, -4//9, -1//2])
3-element Vector{Rational{Int64}}:
 -447//100
  -11//16
  161//72

Main.OurModuleName.compute_Δ — Method

compute_Δ(a, d)

Compute the Δ function from lemma 2 in BGS07.

For a degree $d$ and shift $a > d$, return a vector $[Δ_0,..., Δ_d]$, where $Δ_i = \prod_{j = 0}^d (a + i - j)$.

Example

julia> Main.OurModuleName.compute_Δ(5, 3)
4-element Vector{Integer}:
  120
  360
  840
 1680

Main.OurModuleName.compute_δ — Method

compute_δ(d)

Compute the δ function from lemma 2 in BGS07.

For a degree $d$, return a vector $[δ_0,..., δ_d]$, where $δ_i = \prod_{j = 0, j \neq i}^d (i - j)$.

Example

julia> Main.OurModuleName.compute_δ(5)
6-element Vector{Rational}:
 -1//120
  1//24
 -1//12
  1//12
 -1//24
  1//120

Main.OurModuleName.lagrange — Method

lagrange(p, a)

Lagrange interpolate a vector $p$ of polynomial outputs to a new position $a$.

For a polynomial $P$, apply Lagrange interpolation to the values $p = [P(0), P(1), ..., P(d)]$ to produce a vector of shifted values $[P(a), P(a + 1), ..., P(a + d)]$.

Example

julia> Main.OurModuleName.lagrange([1//4, -7//4, -31//4], 3)
3-element Vector{Rational{Int64}}:
  -71//4
 -127//4
 -199//4

Main.OurModuleName.lagrange_matrix — Method

lagrange_matrix(A)

Lagrange interpolate all entries of a list of $d+1$ matrices and return $3(d+1)$ matrices

Apply the following to each entry of the matrices, with $a = d+1, 2d+1, 3d+1$: For a polynomial $P$, apply Lagrange interpolation to the values $p = [P(0), P(1), ..., P(d)]$ to produce a vector of shifted values $[P(a), P(a + 1), ..., P(a + d)]$.

Example

julia> Main.OurModuleName.lagrange_matrix([0:2 3:5 6:8 ;;; 9:11 12:14 15:17])
3×3×6 Array{Int64, 3}:
[:, :, 1] =
 18  21  24
 19  22  25
 20  23  26

[:, :, 2] =
 27  30  33
 28  31  34
 29  32  35

[:, :, 3] =
 36  39  42
 37  40  43
 38  41  44

[:, :, 4] =
 45  48  51
 46  49  52
 47  50  53

[:, :, 5] =
 54  57  60
 55  58  61
 56  59  62

[:, :, 6] =
 63  66  69
 64  67  70
 65  68  71